pyproject.toml
[project]
name = "resourceful"
version = "0.1.0"
description = "Home resource broker: monetize excess solar, bandwidth, and GPU compute"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "PyJWT>=2.7.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.2",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.4",
    "httpx>=0.27.0",
    "stripe>=10.0.0",
    "itsdangerous>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]
prod = [
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
